chore: move ledger app check into a hook
